requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
bcrypt==4.0.1
billiard==3.6.4.0
blinker==1.4
Brotli==1.0.9
cachelib==0.9.0
celery==5.2.7
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==22.4.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect2==0.7.0
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.3
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.38.0
greenlet==2.0.1
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==5.1.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.1.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
passlib==1.7.4
Pillow==9.4.0
prompt-toolkit==3.0.36
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydyf==0.5.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyphen==0.13.2
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.6
PyYAML==5.3.1
redis==4.4.0
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.4.44
ssh-import-id==5.10
systemd-python==234
tinycss2==1.2.1
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
weasyprint==57.2
webencodings==0.5.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==1.0.0
zope.interface==4.7.1
zopfli==0.2.2
